kora integration

+++ application/controllers/auth.go
@@ -711,15 +711,21 @@ func VerifyAccount(ctx *interfaces.ApplicationContext[dto.VerifyAccountData]) {
 		},
 		"tier": 1,
 	}
-	userRepo.UpdatePartialByFilter(map[string]interface{}{
+	success, err = userRepo.UpdatePartialByFilter(map[string]interface{}{
 		"email": ctx.GetStringContextData("Email"),
 	}, userUpdatedInfo)
 	if err != nil {
 		apperrors.UnknownError(ctx.Ctx, err, ctx.GetHeader("Polymer-Device-Id"))
 		return
 	}
+	if !success {
+		apperrors.UnknownError(ctx.Ctx, err, ctx.GetHeader("Polymer-Device-Id"))
+		return
+	}
 	if ctx.Body.Path == "bvn" {
-		wallet.GenerateNGNDVA(ctx.Ctx, account.WalletID, ctx.GetStringContextData("FirstName"), ctx.GetStringContextData("LastName"), ctx.GetStringContextData("Email"), ctx.Body.ID, ctx.GetHeader("Polymer-Device-Id"))
+		wallet.GenerateNGNDVA(ctx.Ctx, account.WalletID, kycDetails.FirstName, kycDetails.LastName, ctx.GetStringContextData("Email"), ctx.Body.ID, *utils.GenerateDummyKYCID(), ctx.GetHeader("Polymer-Device-Id"))
+	} else {
+		wallet.GenerateNGNDVA(ctx.Ctx, account.WalletID, kycDetails.FirstName, kycDetails.LastName, ctx.GetStringContextData("Email"), *utils.GenerateDummyKYCID(), ctx.Body.ID, ctx.GetHeader("Polymer-Device-Id"))
 	}
 	cache.Cache.DeleteOne(fmt.Sprintf("%s-kyc-attempts-left", ctx.GetStringContextData("Email")))
 	cache.Cache.DeleteOne(fmt.Sprintf("%s-%s-kyc-attempts-left", ctx.GetStringContextData("Email"), ctx.Body.Path))
